Integrate refactors into branch
